backend/database/supabase/requirements.txt
# Database connection requirements
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
dj-database-url>=2.1.0
supabase>=0.7.1